+++ faria.js
@@ -260,6 +260,9 @@ faria = {
       return desc
         ? array.sort( (a, b) => (b[1] - a[1]) )
         : array.sort( (a, b) => (a[1] - b[1]) );
+    },
+    deepIncludes: function ( array, target ) {
+      return JSON.stringify(array).includes(JSON.stringify(target));
     }
   }
 };